+++ src/core/prelude/string/const.ts
@@ -14,5 +14,13 @@ export const
 
 export const
 	isDigital = /\d/,
+	isConcatChar = /\u200D/,
+	isFlagLetters = /\uD83C[\uDDE6-\uDDFF]/;
+
+export const
+	// eslint-disable-next-line no-misleading-character-class
+	isCombinable = /[\u0300-\u036F\u1AB0-\u1AFF\u200D\u20D0-\u20FF\uFE0F]|\uD83C[\uDFFB-\uDFFF]/;
+
+export const
 	normalizeRgxp = /(^[\s_-]+)|([\s_-]+$)|([\s_-]+)/g,
 	camelizeRgxp = /(^[\s_-]+)|([\s_-]+$)|[\s_-]+([^\s-]|$)/g;